requirements.txt
# 百家樂自動投注機器人依賴套件

# 核心 GUI 框架
PySide6>=6.5.0

# 圖像處理
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# 自動化操作
pyautogui>=0.9.54

# HTTP 客戶端與服務器
requests>=2.31.0
flask>=2.3.0

# 環境變數管理
python-dotenv>=1.0.0

# YAML 配置解析
pyyaml>=6.0

# 日誌與時間處理
python-dateutil>=2.8.2

# 數據處理
pandas>=2.0.0

# 可選：開發和測試工具
pytest>=7.4.0
pytest-qt>=4.2.0
black>=23.0.0
flake8>=6.0.0

# 可選：進階圖像處理
scikit-image>=0.21.0